move ImagePyramid to cudalegacy
